separate windows make slo structure and sorter
for some reason the try portion of slo sorter efficient didnt throw an
error when it hit the first slo file, so it turned it into
"reflectance", overwriting the old file instead of jumping to the except
and creating the folder.

+++ sorters/SLO_sorter_efficient_win.py
@@ -8,29 +8,29 @@ sp = os.sep
 def sorting(source):
         file_list = os.listdir(source)
         print (source)
-        destination_apd1 = os.join(source, 'reflectance')+sp #the trailing / is very important to allow the try/except statements to work. without it, files are sequentially overwritten
-        destination_apd2 = os.join(source,'fluorescence')+sp
+        destination_apd1 = os.path.join(source, 'reflectance')
+        destination_apd2 = os.path.join(source,'fluorescence')
 
         for each_file in file_list:
-                file_path = os.join(source,each_file)
+                file_path = os.path.join(source,each_file)
                 if os.path.isfile(file_path) is True:
                         if 'APD1' in each_file:
-                                try:
-                                    shutil.move(file_path, destination_apd1)       
-                                except:
+                                if not os.path.exists(destination_apd1):
                                     os.makedirs(destination_apd1)
                                     shutil.move(file_path, destination_apd1)       
+                                else:
+                                    shutil.move(file_path, destination_apd1)       
                         elif 'APD2' in each_file:
-                                try:
-                                    shutil.move(file_path, destination_apd2)
-                                except:
+                                if not os.path.exists(destination_apd2):
                                     os.makedirs(destination_apd2)
                                     shutil.move(file_path, destination_apd2)
+                                else:
+                                    shutil.move(file_path, destination_apd2)
                 elif os.path.isdir(file_path) is True:
-                        if file_path == os.join(source,'fluorescence') or file_path == os.join(source,'reflectance'):
+                        if file_path == os.path.join(source,'fluorescence') or file_path == os.path.join(source,'reflectance'):
                                 continue
                         else:
-                                new_source = os.join(source,each_file)
+                                new_source = os.path.join(source,each_file)
                                 sorting(new_source)
 
 def main():
